# Merge index (based on chaining index implementation)