@aawsome 2020-07-04
Here are benchmark results, where old is based on commit bb65143.

```
name                                         old time/op    new time/op    delta
PackerManager-4                                 265ms ± 2%     267ms ± 3%     ~     (p=0.113 n=9+10)
DecodeIndex-4                                  22.3µs ± 4%    22.4µs ± 2%     ~     (p=0.353 n=10+10)
IndexHasUnknown-4                              74.4ns ±17%    77.2ns ±13%     ~     (p=0.342 n=10+10)
IndexHasKnown-4                                72.0ns ± 4%    73.9ns ±11%     ~     (p=0.200 n=9+10)
IndexAlloc-4                                    900ms ± 2%     891ms ± 5%     ~     (p=0.218 n=10+10)
MasterIndexLookupSingleIndex-4                  229ns ± 6%     259ns ± 3%  +13.14%  (p=0.000 n=10+10)
MasterIndexLookupMultipleIndex-4                511ns ± 6%     252ns ± 7%  -50.60%  (p=0.000 n=9+10)
MasterIndexLookupSingleIndexUnknown-4          97.8ns ±10%   125.2ns ±13%  +28.07%  (p=0.000 n=10+10)
MasterIndexLookupMultipleIndexUnknown-4         420ns ±11%     101ns ± 9%  -75.92%  (p=0.000 n=10+10)
MasterIndexLookupMultipleIndexReal-4           10.5µs ± 6%     0.3µs ± 7%  -97.52%  (p=0.000 n=10+9)
MasterIndexLookupMultipleIndexRealUnknown-4    10.2µs ± 4%     0.1µs ±11%  -99.02%  (p=0.000 n=10+10)
SaveAndEncrypt-4                                136ns ± 2%     137ns ± 2%     ~     (p=0.230 n=9+10)
LoadTree-4                                      442µs ± 2%     467µs ± 1%   +5.78%  (p=0.000 n=10+10)
LoadBlob-4                                     8.23ms ± 0%    8.34ms ± 1%   +1.36%  (p=0.000 n=9+10)
LoadAndDecrypt-4                               10.2ms ± 2%    10.2ms ± 2%     ~     (p=0.436 n=9+9)
LoadIndex-4                                    43.8ms ± 3%    43.5ms ± 4%     ~     (p=0.631 n=10+10)

name                                         old speed      new speed      delta
PackerManager-4                               199MB/s ± 2%   197MB/s ± 3%     ~     (p=0.108 n=9+10)
SaveAndEncrypt-4                             30.9TB/s ± 2%  30.7TB/s ± 2%     ~     (p=0.243 n=9+10)
LoadBlob-4                                    121MB/s ± 0%   120MB/s ± 1%   -1.34%  (p=0.000 n=9+10)
LoadAndDecrypt-4                             98.5MB/s ± 2%  98.1MB/s ± 2%     ~     (p=0.449 n=9+9)

name                                         old alloc/op   new alloc/op   delta
PackerManager-4                                91.0kB ± 0%    91.0kB ± 0%     ~     (p=0.510 n=9+10)
IndexAlloc-4                                    400MB ± 0%     400MB ± 0%     ~     (p=0.579 n=10+10)
SaveAndEncrypt-4                                2.00B ± 0%     2.00B ± 0%     ~     (all equal)

name                                         old allocs/op  new allocs/op  delta
PackerManager-4                                 1.41k ± 0%     1.41k ± 0%     ~     (p=1.000 n=10+10)
IndexAlloc-4                                    1.12M ± 0%     1.12M ± 0%     ~     (p=0.516 n=10+10)
SaveAndEncrypt-4                                 0.00           0.00          ~     (all equal)
```