@aawsome 2020-07-04
Also checked test coverage for the changes due to this PR: only the `grow()` in `moveto` is not called in tests, but is called in the benchmarks. All other changes are covered by tests.