@aawsome 2020-07-17
I rebased to the current version of #2812 and fixed the issues @greatroar found.