@aawsome 2020-07-18
Now that this PR only calls `indexMap.add()` directly, the benchmarks need an update - coming soon.