@aawsome 2020-07-19
The benchmarks got killed on my 8GB laptop due to memory problems; however, the important benchmarks did run. Here is the result comparing old = 93c3601 with new = 0daff59 :
```
name                                             old time/op    new time/op    delta
IndexMapHash-4                                     5.88µs ± 2%    5.87µs ± 2%     ~     (p=0.837 n=10+10)
PackerManager-4                                     270ms ± 1%     267ms ± 1%   -0.91%  (p=0.046 n=9+8)
DecodeIndex-4                                       5.45s ± 1%     5.46s ± 2%     ~     (p=0.853 n=10+10)
DecodeIndexParallel-4                               5.43s ± 0%     5.39s ± 0%   -0.80%  (p=0.000 n=9+7)
IndexHasUnknown-4                                  76.1ns ±12%    73.5ns ±16%     ~     (p=0.342 n=10+10)
IndexHasKnown-4                                    76.4ns ±10%    74.8ns ± 9%     ~     (p=0.271 n=10+10)
IndexAlloc-4                                        848ms ± 1%     847ms ± 1%     ~     (p=1.000 n=8+10)
IndexAllocParallel-4                                513ms ± 3%     514ms ± 2%     ~     (p=0.666 n=9+9)
MasterIndexLookupSingleIndex-4                      198ns ± 6%     192ns ± 7%     ~     (p=0.138 n=9+10)
MasterIndexLookupMultipleIndex-4                   10.3µs ± 3%     0.3µs ±13%  -97.51%  (p=0.000 n=10+10)
MasterIndexLookupSingleIndexUnknown-4               104ns ±16%     103ns ±15%     ~     (p=0.809 n=10+10)
MasterIndexLookupMultipleIndexUnknown-4            10.0µs ± 3%     0.1µs ± 6%  -99.06%  (p=0.000 n=10+9)
MasterIndexLookupParallel/known,indices=100-4      4.89µs ± 1%    4.89µs ± 1%     ~     (p=0.868 n=10+10)
MasterIndexLookupParallel/unknown,indices=100-4    4.55µs ± 1%    4.64µs ± 2%   +2.16%  (p=0.000 n=8+9)

name                                             old speed      new speed      delta
IndexMapHash-4                                    697MB/s ± 2%   698MB/s ± 2%     ~     (p=0.853 n=10+10)
PackerManager-4                                   195MB/s ± 1%   197MB/s ± 1%   +0.91%  (p=0.046 n=9+8)

name                                             old alloc/op   new alloc/op   delta
IndexMapHash-4                                      0.00B          0.00B          ~     (all equal)
PackerManager-4                                    91.0kB ± 0%    91.0kB ± 0%     ~     (p=0.714 n=10+10)
IndexAlloc-4                                        401MB ± 0%     401MB ± 0%     ~     (all equal)
IndexAllocParallel-4                                400MB ± 0%     400MB ± 0%     ~     (all equal)

name                                             old allocs/op  new allocs/op  delta
IndexMapHash-4                                       0.00           0.00          ~     (all equal)
PackerManager-4                                     1.41k ± 0%     1.41k ± 0%     ~     (p=0.714 n=10+10)
IndexAlloc-4                                         907k ± 0%      907k ± 0%     ~     (all equal)
IndexAllocParallel-4                                 907k ± 0%      907k ± 0%     ~     (all equal)
```