@MichaelEischer 2020-07-20
> The benchmarks got killed on my 8GB laptop due to unsufficient memory; however, the important benchmarks did run.

Insufficient memory on a laptop with 8GB RAM sounds a bit worrying, as in that case either the tests use too large indexes or have a memory leak. Which test failed exactly? FWIW the benchmarks completed successfully on my laptop and that also has "just" 8GB RAM.