@aawsome 2020-07-21
> > The benchmarks got killed on my 8GB laptop due to unsufficient memory; however, the important benchmarks did run.
> 
> Insufficient memory on a laptop with 8GB RAM sounds a bit worrying, as in that case either the tests use too large indexes or have a memory leak. Which test failed exactly? FWIW the benchmarks completed successfully on my laptop and that also has "just" 8GB RAM.

I had some other programs in memory which also took their share...

The memory-hungry benchmark was `MasterIndexLookupParallel` which simulates 1M, 2M and 4M pack files (the other tests had a maximum of ~1M pack files). I changed it such that it simulates 250k, 500k and 1M pack files.

Here are the benchmark results:
```
name                                             old time/op    new time/op    delta
IndexMapHash-4                                     5.77µs ± 2%    5.90µs ± 2%   +2.38%  (p=0.000 n=10+10)
PackerManager-4                                     267ms ± 2%     267ms ± 1%     ~     (p=0.720 n=10+9)
DecodeIndex-4                                       5.39s ± 0%     5.43s ± 1%   +0.75%  (p=0.015 n=8+9)
DecodeIndexParallel-4                               5.39s ± 1%     5.41s ± 1%   +0.40%  (p=0.019 n=10+10)
IndexHasUnknown-4                                  79.0ns ±26%    73.3ns ±12%     ~     (p=0.243 n=10+9)
IndexHasKnown-4                                    75.3ns ±10%    75.6ns ±15%     ~     (p=0.897 n=10+10)
IndexAlloc-4                                        837ms ± 1%     890ms ± 5%   +6.32%  (p=0.000 n=9+10)
IndexAllocParallel-4                                450ms ± 3%     511ms ± 1%  +13.54%  (p=0.000 n=9+9)
MasterIndexLookupSingleIndex-4                      203ns ± 5%     196ns ± 5%   -3.60%  (p=0.030 n=10+10)
MasterIndexLookupMultipleIndex-4                   10.1µs ± 5%     0.2µs ±10%  -97.54%  (p=0.000 n=10+10)
MasterIndexLookupSingleIndexUnknown-4               100ns ±11%     104ns ± 9%     ~     (p=0.092 n=10+10)
MasterIndexLookupMultipleIndexUnknown-4            9.94µs ± 5%    0.09µs ± 7%  -99.06%  (p=0.000 n=10+9)
MasterIndexLookupParallel/known,indices=25-4       1.50µs ± 2%    0.44µs ± 2%  -70.43%  (p=0.000 n=10+10)
MasterIndexLookupParallel/unknown,indices=25-4     1.06µs ± 1%    0.30µs ± 1%  -71.17%  (p=0.000 n=10+9)
MasterIndexLookupParallel/known,indices=50-4       2.45µs ± 1%    0.47µs ± 3%  -80.73%  (p=0.000 n=9+9)
MasterIndexLookupParallel/unknown,indices=50-4     2.14µs ± 1%    0.30µs ± 6%  -85.90%  (p=0.000 n=10+10)
MasterIndexLookupParallel/known,indices=100-4      4.92µs ± 0%    0.46µs ± 5%  -90.57%  (p=0.000 n=8+10)
MasterIndexLookupParallel/unknown,indices=100-4    4.71µs ± 1%    0.31µs ± 2%  -93.39%  (p=0.000 n=9+10)
SaveAndEncrypt-4                                   42.7ms ± 1%    43.6ms ± 1%   +2.06%  (p=0.000 n=9+10)
LoadTree-4                                          437µs ± 3%     441µs ± 2%     ~     (p=0.065 n=9+10)
LoadBlob-4                                         8.22ms ± 0%    8.43ms ± 1%   +2.54%  (p=0.000 n=8+9)
LoadAndDecrypt-4                                   8.58ms ± 1%    8.58ms ± 0%     ~     (p=0.356 n=10+9)
LoadIndex-4                                        32.7ms ± 1%    33.6ms ± 1%   +2.93%  (p=0.000 n=10+10)

name                                             old speed      new speed      delta
IndexMapHash-4                                    710MB/s ± 2%   694MB/s ± 2%   -2.33%  (p=0.000 n=10+10)
PackerManager-4                                   197MB/s ± 2%   198MB/s ± 1%     ~     (p=0.720 n=10+9)
SaveAndEncrypt-4                                 98.2MB/s ± 1%  96.3MB/s ± 1%   -2.02%  (p=0.000 n=9+10)
LoadBlob-4                                        122MB/s ± 0%   119MB/s ± 1%   -2.48%  (p=0.000 n=8+9)
LoadAndDecrypt-4                                  117MB/s ± 1%   117MB/s ± 0%     ~     (p=0.344 n=10+9)

name                                             old alloc/op   new alloc/op   delta
IndexMapHash-4                                      0.00B          0.00B          ~     (all equal)
PackerManager-4                                    91.1kB ± 0%    91.1kB ± 0%     ~     (all equal)
IndexAlloc-4                                        401MB ± 0%     401MB ± 0%     ~     (all equal)
IndexAllocParallel-4                                400MB ± 0%     400MB ± 0%   +0.00%  (p=0.000 n=9+10)
SaveAndEncrypt-4                                   21.0MB ± 0%    21.0MB ± 0%   +0.00%  (p=0.000 n=10+9)

name                                             old allocs/op  new allocs/op  delta
IndexMapHash-4                                       0.00           0.00          ~     (all equal)
PackerManager-4                                     1.41k ± 0%     1.41k ± 0%     ~     (all equal)
IndexAlloc-4                                         907k ± 0%      907k ± 0%     ~     (all equal)
IndexAllocParallel-4                                 907k ± 0%      907k ± 0%   +0.00%  (p=0.000 n=9+10)
SaveAndEncrypt-4                                      118 ± 0%       118 ± 0%     ~     (all equal)
```
(wouldn't count too much about the ~2% differences, as said this is my laptop and there are other things like a firefox running...)